Fixing the registration handle and communication
