+++ src/components/LogReg.svelte
@@ -7,10 +7,10 @@
 		window.location.href = baseUrl + "/home";
 	};
 	let change = "Registration";
-	let login = true;
+	let loginer = true;
 	function handleClick(event) {
-		login = !login;
-		if (login == true){ 
+		loginer = !loginer;
+		if (loginer == true){ 
 			change = "Registration";
 		}
 		else {
@@ -18,16 +18,14 @@
 		}
 	}
 </script>
-
 <link rel="icon" type="image/png" sizes="500x500" href="assets/img/logo.png" />
 <link rel="stylesheet" href="assets/css/styles.css" />
 <link rel="stylesheet" href="assets/css/Login-Form-Dark.css" />
 <section class="login-dark">
-	{#if login}
+	{#if loginer}
 		<Login />
 	{:else}
 		<Register />
 	{/if}
-	<button type="button" on:click={handleClick} class="{login ? 'btn btn-primary d-block changer':'btn btn-primary d-block changer1'}"
-		>{change}</button>
+	<button type="button" on:click={handleClick} class="{loginer ? 'btn btn-primary d-block changer':'btn btn-primary d-block changer1'}">{change}</button>
 </section>
